requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
google-generativeai
faiss-cpu
langchain-google-genai